+++ kotlin-source/base/src/main/kotlin/net/corda/examples/oracle/flow/QueryPrime.kt
@@ -5,10 +5,9 @@ import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.InitiatingFlow
 import net.corda.core.identity.Party
 import net.corda.core.utilities.unwrap
-import java.math.BigInteger
 
 // Simple flow which takes a reference to an Oracle and a number then returns the corresponding nth prime number.
 @InitiatingFlow
-class QueryPrime(val oracle: Party, val n: Int) : FlowLogic<BigInteger>() {
-    @Suspendable override fun call() = sendAndReceive<BigInteger>(oracle, n).unwrap { it }
+class QueryPrime(val oracle: Party, val n: Int) : FlowLogic<Int>() {
+    @Suspendable override fun call() = sendAndReceive<Int>(oracle, n).unwrap { it }
 }
